Changed URI prefix of git-server from git to git-server
Previously the URI of git server started with "/git/" which is ambigious with git APIs of Che. This situation is ok when git-server is packaged as separate webapp and handles all requests (serve("/*")), but if the git-server is packaged as a part of api.war this servlet needs unique URI mapping. This change has a corresponding pull request with the same name in che-plugins.